+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ### Ver 0.1.1(Not released)
 
   * #1 [Improve tab title color](https://github.com/georgeOsdDev/react-draggable-tab/issues/1)
+  * #2 [onTabSelected, onTabClosed, onTabPositionChanged should be called in setState callback](https://github.com/georgeOsdDev/react-draggable-tab/issues/2)
 
 ### Ver 0.1.0 Initial release