+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Change Log
 
-### Ver 0.1.1(Not released)
+### Ver 0.2.1(Next release, Not released)
+
+### Ver 0.2.0
 
   * #1 [Improve tab title color](https://github.com/georgeOsdDev/react-draggable-tab/issues/1)
   * #2 [onTabSelected, onTabClosed, onTabPositionChanged should be called in setState callback](https://github.com/georgeOsdDev/react-draggable-tab/issues/2)